Moved alias transformations to a concern
